fix: update session type in session_results

+++ src/main/java/com/arctic/backend_for_arctic_team/metrics/model/dto/request/SessionDto.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public record SessionDto(
         @JsonProperty("individualNumber") String individualNumber,
         @JsonProperty("expeditionId") String expeditionId,
-        @JsonProperty("session") Long session,
+        @JsonProperty("session") Integer session,
         @JsonProperty("objectiveCognitive") Integer objectiveCognitive,
         @JsonProperty("objectivePsychological") Integer objectivePsychological,
         @JsonProperty("objectivePhysiological") Integer objectivePhysiological,
